Move file structure around in betterttv resolver
