+++ internal/resolvers/betterttv/emote_loader_test.go
