--- internal/resolvers/betterttv/static_responses.go
@@ -1,14 +1,0 @@
-package betterttv
-
-import (
-	"net/http"
-
-	"github.com/Chatterino/api/pkg/resolver"
-)
-
-var (
-	emoteNotFoundResponse = &resolver.Response{
-		Status:  http.StatusNotFound,
-		Message: "No BetterTTV emote with this hash found",
-	}
-)
